`setup-fork`: improve upstream remote heuristics
If specified upstream remote does not exist, try to iterate through all
remotes and pick one that has a GitHub URL. Do not blindly pick the only
existing remote as it may be wrong in case of user error.

Additionally, validate that a remote has a GitHub URL before attempting
to rename it.