`setup-fork`: do not print the "Nothing to do, remote XXX ..." message
If we are coming from `git hub setup-fork` (i. e. the repo was not just
cloned), hide the "Nothing to do. remote XXX already exists" message
as it might be confusing for the user (especially that it may refer to
a newly-renamed remote).

It is not clear to me why this logic even exists (naturally, if we are
coming from `git hub setup-fork`, then the remote that we've just picked
will always exist), but I'm not removing it as it might serve some other
purpose.